Remove deprecated storybook usages
