+++ packages/datepickers/stories/examples/CustomFormat.tsx
@@ -6,16 +6,11 @@
  */
 
 import React, { useMemo, useState } from 'react';
-import { Meta, Story } from '@storybook/react';
+import { Story } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { Datepicker } from '@zendeskgarden/react-datepickers';
 import { Grid, Row, Col } from '@zendeskgarden/react-grid';
 import { Field, Label, Hint, Input } from '@zendeskgarden/react-forms';
-
-export default {
-  title: 'Components/Datepickers/Datepicker',
-  subcomponents: { Datepicker }
-} as Meta;
 
 /**
  * https://github.com/storybookjs/storybook/issues/13362
@@ -61,7 +56,8 @@ export const CustomFormat: Story = ({ foo }) => {
 
 CustomFormat.parameters = {
   docs: {
-    storyDescription: `
+    description: {
+      story: `
 By default Garden use the [Intl.DateTimeFormat](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DateTimeFormat)
 localization utility to format our dates.
 
@@ -69,5 +65,6 @@ This logic is customizable with the \`formatDate\` prop where you may use
 other date formatting libraries like [momentjs](https://momentjs.com/) or
 [date-fns](https://date-fns.org/).
       `
+    }
   }
 };
